requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
numpy==1.26.4
tensorflow-cpu==2.15.0
gunicorn==21.2.0
